-corretta tipologia parametro spr_GetWorkoutPlans -aggiunto parametro per gestione archiviazione workout plans -corretta comunicazione id con il database in quanto mssql non supporta ulong(long/ulong) -aggiunta chiamata per cancellare le serie da un workout

+++ bb-project.Server/bb-project.DB/bb-project.DB/StoredProcedures/spr_GetWorkoutPlans.sql
@@ -1,5 +1,5 @@
 ﻿CREATE PROCEDURE [dbo].[spr_GetWorkoutPlans]
-    @workoutPlanId int = 0,
+    @workoutPlanId BIGINT = 0,
     @getArchived BIT = 0
 AS
     BEGIN
